Added Token Authentification and made OpenApi more detailed

+++ src/main/java/com/yorastd/projectmanagement/Repositories/TaskNodeRepo.java
@@ -1,9 +1,9 @@
 package com.yorastd.projectmanagement.Repositories;
 
-import com.yorastd.projectmanagement.Models.TaskTreeModel.TaskNode;
+import com.yorastd.projectmanagement.Models.Tasks.TaskTreeModel.TaskNode;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskNodeRepo extends JpaRepository<TaskNode,Integer> {
+public interface TaskNodeRepo extends JpaRepository<TaskNode,Long> {
 }
